+++ src/test/java/org/opensearch/flowframework/workflow/UndeployModelStepTests.java
@@ -72,7 +72,8 @@ public class UndeployModelStepTests extends OpenSearchTestCase {
             inputData.getNodeId(),
             inputData,
             Map.of("step_1", new WorkflowData(Map.of(MODEL_ID, modelId), "workflowId", "nodeId")),
-            Map.of("step_1", MODEL_ID)
+            Map.of("step_1", MODEL_ID),
+            Collections.emptyMap()
         );
         verify(machineLearningNodeClient).undeploy(any(String[].class), any(), any());
 
@@ -86,6 +87,7 @@ public class UndeployModelStepTests extends OpenSearchTestCase {
         PlainActionFuture<WorkflowData> future = UndeployModelStep.execute(
             inputData.getNodeId(),
             inputData,
+            Collections.emptyMap(),
             Collections.emptyMap(),
             Collections.emptyMap()
         );
@@ -118,7 +120,8 @@ public class UndeployModelStepTests extends OpenSearchTestCase {
             inputData.getNodeId(),
             inputData,
             Map.of("step_1", new WorkflowData(Map.of(MODEL_ID, "test"), "workflowId", "nodeId")),
-            Map.of("step_1", MODEL_ID)
+            Map.of("step_1", MODEL_ID),
+            Collections.emptyMap()
         );
 
         verify(machineLearningNodeClient).undeploy(any(String[].class), any(), any());